# docs/swagger.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mini E-Commerce API",
    "version": "v1"
  },
  "paths": {
    "/api/AdminOrdenes": {
      "get": {
        "tags": [
          "AdminOrdenes"
        ],
        "summary": "Obtiene todas las órdenes del sistema.",
        "description": "Requiere rol <b>Admin</b>.\r\nDevuelve todas las órdenes de todos los usuarios.",
        "responses": {
          "200": {
            "description": "Listado de órdenes.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "No tiene rol Admin.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/AdminOrdenes/ordenes-usuario/{usuarioId}": {
      "get": {
        "tags": [
          "AdminOrdenes"
        ],
        "summary": "Obtiene todas las órdenes de un usuario específico.",
        "description": "Requiere rol <b>Admin</b>.\r\nValida que el usuario exista.",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "description": "Id del usuario.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de órdenes del usuario.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "usuarioId inválido.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "No tiene rol Admin.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/AdminOrdenes/orden-detalles/{ordenId}": {
      "get": {
        "tags": [
          "AdminOrdenes"
        ],
        "summary": "Obtiene el detalle de una orden por Id (modo Admin).",
        "description": "Requiere rol <b>Admin</b>.\r\nDevuelve el detalle de la orden aunque no pertenezca al Admin.",
        "parameters": [
          {
            "name": "ordenId",
            "in": "path",
            "description": "Id de la orden.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la orden.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ordenId inválido.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Orden no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "No tiene rol Admin.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Ordenes/{ordenId}": {
      "get": {
        "tags": [
          "Ordenes"
        ],
        "summary": "Obtiene el detalle de una orden específica del usuario.",
        "description": "Requiere autenticación.\r\nSolo devuelve la orden si pertenece al usuario autenticado.",
        "parameters": [
          {
            "name": "ordenId",
            "in": "path",
            "description": "Id de la orden.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la orden.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ordenId inválido o usuarioId inválido en token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Orden no existe o no pertenece al usuario.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ObjectApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
      "OrdenDetallesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OrdenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/StatusOrden"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "ordenDetallesDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenDetallesDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrdenDtoApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "$ref": "#/components/schemas/OrdenDto"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
      "OrdenDtoListApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenDto"
            },
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "StatusOrden": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese: Bearer {tu_token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}